feat(图表): 明细表支持总计 #12650
